requirements.txt
djangorestframework==3.14.0
Django==4.2.4
